Criando endpoint de atualizar previsões de ia por id do produto

+++ src/ssmai_backend/schemas/ai_analysis_schemas.py
@@ -15,7 +15,7 @@ class HistoricoItem(BaseModel):
 
 class PrevisaoItem(BaseModel):
     data: datetime
-    estoque_previsto: float
+    estoque_previsto: int
 
 class PrevisoesResponse(BaseModel):
     historico: list[HistoricoItem]
